+++ app/(tabs)/perfil/index.tsx
@@ -1,32 +1,102 @@
 import { useEffect, useState } from 'react';
-import { View, Text, Image, StyleSheet, Button, TextInput, ActivityIndicator } from 'react-native';
+import {
+  View,
+  Text,
+  Image,
+  StyleSheet,
+  Button,
+  TextInput,
+  TouchableOpacity,
+  ActivityIndicator,
+} from 'react-native';
 import { Stack, router } from 'expo-router';
 import * as ImagePicker from 'expo-image-picker';
-import { getAuth } from 'firebase/auth';
-import { doc, getDoc, updateDoc } from 'firebase/firestore';
+import {
+  getAuth,
+  onAuthStateChanged,
+  signInWithEmailAndPassword,
+  createUserWithEmailAndPassword,
+} from 'firebase/auth';
+import { doc, getDoc, setDoc, updateDoc } from 'firebase/firestore';
 import { ref, deleteObject, uploadBytes, getDownloadURL } from 'firebase/storage';
-import { db, storage } from '../../../firebase';
+import { app, db, storage } from '../../../firebase';
 
 export default function HomeScreen() {
-  const auth = getAuth();
-  const user = auth.currentUser;
+  const auth = getAuth(app);
+  const [user, setUser] = useState(auth.currentUser);
   const [userData, setUserData] = useState<any>(null);
+
+  const [email, setEmail] = useState('');
+  const [password, setPassword] = useState('');
+  const [username, setUsername] = useState('');
+  const [edad, setEdad] = useState('');
+  const [foto, setFoto] = useState<ImagePicker.ImagePickerAsset | null>(null);
+  const [isRegistering, setIsRegistering] = useState(false);
+
   const [message, setMessage] = useState('');
   const [isError, setIsError] = useState(false);
   const [loading, setLoading] = useState(false);
 
   useEffect(() => {
-    const cargarDatos = async () => {
-      if (user) {
-        const docRef = doc(db, 'usuarios', user.uid);
+    const unsubscribe = onAuthStateChanged(auth, async (u) => {
+      setUser(u);
+      if (u) {
+        const docRef = doc(db, 'usuarios', u.uid);
         const snap = await getDoc(docRef);
         if (snap.exists()) {
           setUserData(snap.data());
         }
+      } else {
+        setUserData(null);
       }
-    };
-    cargarDatos();
-  }, [user]);
+    });
+    return unsubscribe;
+  }, []);
+
+  const pickImage = async () => {
+    const res = await ImagePicker.launchImageLibraryAsync({
+      mediaTypes: ImagePicker.MediaTypeOptions.Images,
+      quality: 0.5,
+    });
+    if (!res.canceled) {
+      setFoto(res.assets[0]);
+    }
+  };
+
+  const handleAuth = async () => {
+    setMessage('');
+    setIsError(false);
+    setLoading(true);
+    try {
+      if (isRegistering) {
+        if (!foto) {
+          setIsError(true);
+          setMessage('Seleccioná una foto de perfil');
+          setLoading(false);
+          return;
+        }
+        const userCred = await createUserWithEmailAndPassword(auth, email, password);
+        const storageRef = ref(storage, `fotos_perfil/${userCred.user.uid}`);
+        const img = await fetch(foto.uri);
+        const blob = await img.blob();
+        await uploadBytes(storageRef, blob);
+        const fotoURL = await getDownloadURL(storageRef);
+        await setDoc(doc(db, 'usuarios', userCred.user.uid), {
+          email,
+          username,
+          edad: parseInt(edad),
+          fotoURL,
+        });
+      } else {
+        await signInWithEmailAndPassword(auth, email, password);
+      }
+    } catch (err: any) {
+      setIsError(true);
+      setMessage(err.message);
+    } finally {
+      setLoading(false);
+    }
+  };
 
   const pickAndUpload = async () => {
     if (!user) return;
@@ -54,10 +124,65 @@ export default function HomeScreen() {
 
   const logout = async () => {
     await auth.signOut();
-    router.replace('/perfil/login');
   };
 
-  if (!user || !userData) {
+  if (!user) {
+    return (
+      <View style={styles.container}>
+        <Stack.Screen options={{ title: isRegistering ? 'Registrarse' : 'Iniciar sesión' }} />
+        <TextInput
+          style={styles.input}
+          placeholder="Correo"
+          value={email}
+          onChangeText={setEmail}
+          autoCapitalize="none"
+        />
+        <TextInput
+          style={styles.input}
+          placeholder="Contraseña"
+          secureTextEntry
+          value={password}
+          onChangeText={setPassword}
+        />
+        {isRegistering && (
+          <>
+            <TextInput
+              style={styles.input}
+              placeholder="Nombre de usuario"
+              value={username}
+              onChangeText={setUsername}
+            />
+            <TextInput
+              style={styles.input}
+              placeholder="Edad"
+              value={edad}
+              onChangeText={setEdad}
+              keyboardType="numeric"
+            />
+            <TouchableOpacity style={styles.imagePicker} onPress={pickImage}>
+              <Text>{foto ? 'Cambiar foto' : 'Seleccionar foto'}</Text>
+            </TouchableOpacity>
+            {foto && <Image source={{ uri: foto.uri }} style={styles.preview} />}
+          </>
+        )}
+        {message ? (
+          <Text style={{ color: isError ? 'red' : 'green', marginBottom: 10 }}>{message}</Text>
+        ) : null}
+        {loading ? <ActivityIndicator style={{ marginBottom: 10 }} /> : null}
+        <Button title={isRegistering ? 'Crear cuenta' : 'Ingresar'} onPress={handleAuth} />
+        <TouchableOpacity onPress={() => setIsRegistering(!isRegistering)} style={{ marginTop: 10 }}>
+          <Text>{isRegistering ? '¿Ya tienes cuenta? Inicia sesión' : '¿No tienes cuenta? Regístrate'}</Text>
+        </TouchableOpacity>
+        {!isRegistering && (
+          <TouchableOpacity onPress={() => router.push('/forgot')} style={{ marginTop: 10 }}>
+            <Text style={{ color: 'blue' }}>¿Olvidaste tu contraseña?</Text>
+          </TouchableOpacity>
+        )}
+      </View>
+    );
+  }
+
+  if (!userData) {
     return (
       <View style={styles.container}>
         <ActivityIndicator />
@@ -92,6 +217,22 @@ const styles = StyleSheet.create({
     height: 150,
     borderRadius: 75,
     marginVertical: 10,
+  },
+  input: {
+    borderWidth: 1,
+    borderColor: '#ccc',
+    borderRadius: 4,
+    padding: 10,
+    marginBottom: 10,
+    alignSelf: 'stretch',
+  },
+  imagePicker: {
+    padding: 10,
+    backgroundColor: '#ddd',
+    borderRadius: 4,
+    marginBottom: 10,
+    alignItems: 'center',
+    alignSelf: 'stretch',
   },
   title: {
     fontSize: 20,
